Move AwsSpanBuilder to middleware::aws
